pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vox"
version = "0.1.0"
description = "Voice recording pipeline automation"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "PyYAML>=6.0",
    "rich>=13.0",
    "urllib3>=2.0",
]

[project.scripts]
vox = "vox.cli:main"

[project.optional-dependencies]
diarize = [
    "torch>=2.0",
    "torchaudio>=2.0",
    "pyannote-audio>=3.3",
    "speechbrain>=1.0",
    "numpy>=1.24",
]

[tool.setuptools.packages.find]
include = ["vox*"]

[tool.setuptools.package-data]
vox = ["default_analysis_prompt.txt"]
